# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/upload": {
      "post": {
        "operationId": "UploadController_uploadFile",
        "parameters": [
          {
            "name": "folder",
            "required": false,
            "in": "query",
            "description": "Folder name to store the file",
            "schema": {
              "example": "portfolio",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "File to upload (images only)"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "description": "URL of the uploaded file",
                      "example": "https://example.com/uploads/file.jpg"
                    },
                    "filename": {
                      "type": "string",
                      "description": "Name of the uploaded file",
                      "example": "file.jpg"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - File validation failed"
          }
        },
        "summary": "Upload a single file",
        "tags": [
          "upload"
        ]
      }
    },
    "/upload/multiple": {
      "post": {
        "operationId": "UploadController_uploadMultipleFiles",
        "parameters": [
          {
            "name": "folder",
            "required": false,
            "in": "query",
            "description": "Folder name to store the files",
            "schema": {
              "example": "portfolio",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Files to upload (images only, max 10 files)"
                  }
                },
                "required": [
                  "files"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Files uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "url": {
                        "type": "string",
                        "description": "URL of the uploaded file",
                        "example": "https://example.com/uploads/file.jpg"
                      },
                      "filename": {
                        "type": "string",
                        "description": "Name of the uploaded file",
                        "example": "file.jpg"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Files upload failed"
          }
        },
        "summary": "Upload multiple files",
        "tags": [
          "upload"
        ]
      }
    },
...
  },
  "info": {
    "title": "DEPA API",
    "description": "The DEPA API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}